backend/voicebot-engine/requirements.txt
# Pipecat core framework with services
pipecat-ai[daily,azure,openai,google,elevenlabs]>=0.0.50

# Additional dependencies
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
httpx>=0.28.0
asyncpg>=0.30.0
